requirements.txt
anthropic>=0.25.0
ddgs>=9.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
